# Render underwater horizon into mask