--- a/crest/Assets/Crest/Crest/Shaders/Underwater/Resources/UnderwaterMaskHorizon.shader.meta
+++ b/crest/Assets/Crest/Crest/Shaders/Underwater/Resources/UnderwaterMaskHorizon.shader.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: f99e556f3e6e041a48ceb5a41eddae28
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: